Renaming enrtl_submethods to enrtl_parameters
